thông báo xóa kế hoạch tour

+++ QuanLyTourDuLich/src/com/myclass/gui/Application.java
@@ -2736,11 +2736,22 @@ JButton btnHopDong_TimKiem = new JButton("Tìm");
 		btnKeHoachTour_Xoa.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				int selectedRow = tblKeHoachTour.getSelectedRow();
-				String maKeHoach = (String) tblDoan.getValueAt(selectedRow, 0);
-				
-				keHoachTourBUS.deleteById(maKeHoach);
-				keHoachTourTblModel.removeRow(selectedRow);
-				KeHoachTourBUS.listKeHoachTourDTO.remove(selectedRow);
+				if(selectedRow>=0) {
+					int result=JOptionPane.showConfirmDialog(null,"Bạn có chắc muốn xóa thông tin này?","thông báo",JOptionPane.YES_NO_OPTION); 
+					if(result==JOptionPane.YES_OPTION) {
+					String maKeHoach = (String) tblDoan.getValueAt(selectedRow, 0);
+					
+					keHoachTourBUS.deleteById(maKeHoach);
+					keHoachTourTblModel.removeRow(selectedRow);
+					KeHoachTourBUS.listKeHoachTourDTO.remove(selectedRow);
+					}
+					else if(result==JOptionPane.NO_OPTION) {
+						JOptionPane.showMessageDialog(null,"Thông tin này chưa được xóa!");
+					}
+				}
+				if(selectedRow<0 ) {
+					JOptionPane.showMessageDialog(cardQuanLyTaiKhoan, "Bạn chưa chọn trường dữ liệu nào cả");
+				}
 			}
 		});
 		btnKeHoachTour_Xoa.setFont(new Font("Tahoma", Font.PLAIN, 16));
